+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`8663` rotki will now create calendar reminders for claiming bridge withdrawals from Layer 2 networks (Base, Optimism, Arbitrum) after the 7-day waiting period.
 * :feature:`8170` rotki now supports Coinbase Prime.
 * :feature:`8439` rotki will now correctly decode Ethereum/Polygon PoS bridging.
 * :feature:`8809` rotki now correctly tracks bridge transactions between Ethereum and Superchain networks (Base and OP Mainnet).